pyproject.toml
[tool.poetry]
name = "annee-bissextile-api"
version = "0.1.0"
description = ""
authors = ["Fanny Tutenuit <fanny.tutenuit@covage.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
django = "5.1.3"
djangorestframework = "3.15.2"
pytest = "8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
